clippy

+++ crates/nu-command/src/conversions/fill.rs
@@ -166,11 +166,7 @@ fn fill(
         FillAlignment::Left
     };
 
-    let width = if let Some(arg) = width_arg {
-        arg as usize
-    } else {
-        1
-    };
+    let width = if let Some(arg) = width_arg { arg } else { 1 };
 
     let character = if let Some(arg) = character_arg {
         arg
